Figured out a bug causing the decryption to run twice, moved decode methods to decryptable module
