+++ encrypted.txt
@@ -1,1 +1,1 @@
-twebljkpxi
+pgwfwbgizno
